+++ packages/google_maps_places_autocomplete_widgets_native/lib/google_maps_places_autocomplete_widgets_native.dart
@@ -1,0 +1,1 @@
+export 'src/native_place_api_provider.dart';